Split AgentData into multiple files
Also renames io/*.cu to io/*.cpp, as these files need no knowledge of cuda

Also improves AgentFunctionData::operator== (several members were missing)
