+++ src/flamegpu/model/ModelData.cpp
@@ -1,13 +1,12 @@
 #include <iostream>
 
 #include "flamegpu/model/ModelData.h"
-#include "flamegpu/model/EnvironmentDescription.h"
+
+#include "flamegpu/model/AgentData.h"
 #include "flamegpu/model/AgentDescription.h"
 #include "flamegpu/model/AgentFunctionDescription.h"
-#include "flamegpu/model/LayerDescription.h"
-#include "flamegpu/runtime/messaging/BruteForce.h"
-#include "flamegpu/runtime/cuRVE/curve_rtc.h"
-
+#include "flamegpu/model/AgentFunctionData.h"
+#include "flamegpu/model/LayerData.h"
 
 const char *ModelData::DEFAULT_STATE = "default";
 
@@ -20,62 +19,11 @@ ModelData::ModelData(const std::string &model_name)
 
 ModelData::~ModelData() { }
 
-AgentData::AgentData(ModelData *const model, const std::string &agent_name)
-    : initial_state(ModelData::DEFAULT_STATE)
-    , agent_outputs(0)
-    , description(new AgentDescription(model, this))
-    , name(agent_name)
-    , keepDefaultState(false) {
-    states.insert(ModelData::DEFAULT_STATE);
-}
-
-
-AgentFunctionData::AgentFunctionData(std::shared_ptr<AgentData> _parent, const std::string &function_name, AgentFunctionWrapper *agent_function, const std::string &in_type, const std::string &out_type)
-    : func(agent_function)
-    , rtc_source("")
-    , rtc_func_name("")
-    , initial_state(_parent->initial_state)
-    , end_state(_parent->initial_state)
-    , message_output_optional(false)
-    , has_agent_death(false)
-    , condition(nullptr)
-    , rtc_condition_source("")
-    , rtc_func_condition_name("")
-    , parent(_parent)
-    , description(new AgentFunctionDescription(_parent->description->model, this))
-    , name(function_name)
-    , msg_in_type(in_type)
-    , msg_out_type(out_type) { }
-
-AgentFunctionData::AgentFunctionData(std::shared_ptr<AgentData> _parent, const std::string& function_name, const std::string &rtc_function_src, const std::string &in_type, const std::string& out_type, const std::string& code_func_name)
-    : func(0)
-    , rtc_source(rtc_function_src)
-    , rtc_func_name(code_func_name)
-    , initial_state(_parent->initial_state)
-    , end_state(_parent->initial_state)
-    , message_output_optional(false)
-    , has_agent_death(false)
-    , condition(nullptr)
-    , rtc_condition_source("")
-    , rtc_func_condition_name("")
-    , parent(_parent)
-    , description(new AgentFunctionDescription(_parent->description->model, this))
-    , name(function_name)
-    , msg_in_type(in_type)
-    , msg_out_type(out_type) { }
-
-LayerData::LayerData(ModelData *const model, const std::string &layer_name, const ModelData::size_type &layer_index)
-    : description(new LayerDescription(model, this))
-    , name(layer_name)
-    , index(layer_index) { }
-
-/**
- * Copy Constructors
- */
 std::shared_ptr<ModelData> ModelData::clone() const {
-    // Arkwardly cant use shared from this inside constructor, so use raw pts instead
+    // Awkwardly cant use shared from this inside constructor, so use raw pts instead
     return std::shared_ptr<ModelData>(new ModelData(*this));
 }
+
 ModelData::ModelData(const ModelData &other)
     : initFunctions(other.initFunctions)
     , stepFunctions(other.stepFunctions)
@@ -104,89 +52,6 @@ ModelData::ModelData(const ModelData &other)
         layers.push_back(std::shared_ptr<LayerData>(new LayerData(this, *m)));
     }
 }
-
-std::shared_ptr<const AgentData> AgentData::clone() const {
-    std::shared_ptr<AgentData> b = std::shared_ptr<AgentData>(new AgentData(nullptr, *this));
-    // Manually copy construct maps of shared ptr
-    for (const auto f : functions) {
-        b->functions.emplace(f.first, std::shared_ptr<AgentFunctionData>(new AgentFunctionData(nullptr, b, *f.second)));
-    }
-    return b;
-}
-AgentData::AgentData(ModelData *const model, const AgentData &other)
-    : variables(other.variables)
-    , states(other.states)
-    , initial_state(other.initial_state)
-    , agent_outputs(other.agent_outputs)
-    , description(model ? new AgentDescription(model, this) : nullptr)
-    , name(other.name)
-    , keepDefaultState(other.keepDefaultState) { }
-
-AgentFunctionData::AgentFunctionData(ModelData *const model, std::shared_ptr<AgentData> _parent, const AgentFunctionData &other)
-    : func(other.func)
-    , rtc_source(other.rtc_source)
-    , rtc_func_name(other.rtc_func_name)
-    , initial_state(other.initial_state)
-    , end_state(other.end_state)
-    , message_output_optional(other.message_output_optional)
-    , agent_output_state(other.agent_output_state)
-    , has_agent_death(other.has_agent_death)
-    , condition(other.condition)
-    , rtc_condition_source(other.rtc_condition_source)
-    , rtc_func_condition_name(other.rtc_func_condition_name)
-    , parent(_parent)
-    , description(model ? new AgentFunctionDescription(model, this) : nullptr)
-    , name(other.name)
-    , msg_in_type(other.msg_in_type)
-    , msg_out_type(other.msg_out_type) {
-    // Manually perform lookup copies
-    if (model) {
-        if (auto a = other.message_input.lock()) {
-            auto _m = model->messages.find(a->name);
-            if (_m != model->messages.end()) {
-                message_input = _m->second;
-            }
-        } else if (other.msg_in_type != CurveRTCHost::demangle(std::type_index(typeid(MsgNone)))) {
-            THROW InvalidMessageType("Function '%s' is missing bound input message of type '%s', type provided was '%s'.", other.name.c_str(), other.msg_in_type.c_str(), CurveRTCHost::demangle(std::type_index(typeid(MsgNone))).c_str());
-        }
-        if (auto a = other.message_output.lock()) {
-            auto _m = model->messages.find(a->name);
-            if (_m != model->messages.end()) {
-                message_output = _m->second;
-            }
-        } else if (other.msg_out_type != CurveRTCHost::demangle(std::type_index(typeid(MsgNone)))) {
-            THROW InvalidMessageType("Function '%s' is missing bound output message of type '%s'.", other.name.c_str(), other.msg_out_type.c_str(), CurveRTCHost::demangle(std::type_index(typeid(MsgNone))).c_str());
-        }
-        if (auto a = other.agent_output.lock()) {
-            auto _a = model->agents.find(a->name);
-            if (_a != model->agents.end()) {
-                agent_output = _a->second;
-            }
-        }
-    }
-}
-
-LayerData::LayerData(ModelData *const model, const LayerData &other)
-    : host_functions(other.host_functions)
-    , description(model ? new LayerDescription(model, this) : nullptr)
-    , name(other.name)
-    , index(other.index) {
-    // Manually perform lookup copies
-    for (auto _f : other.agent_functions) {
-        for (auto a : model->agents) {
-            for (auto f : a.second->functions) {
-                if (f.second->func == _f->func) {
-                    if (f.second->name == _f->name) {
-                        agent_functions.emplace(f.second);
-                        goto next_agent_fn;
-                    }
-                }
-            }
-        }
-    next_agent_fn: {}
-    }
-}
-
 
 bool ModelData::operator==(const ModelData& rhs) const {
     if (this == &rhs)  // They point to same object
@@ -248,139 +113,7 @@ bool ModelData::operator==(const ModelData& rhs) const {
     }
     return false;
 }
+
 bool ModelData::operator!=(const ModelData& rhs) const {
     return !operator==(rhs);
-}
-bool AgentData::operator==(const AgentData& rhs) const {
-    if (this == &rhs)  // They point to same object
-        return true;
-    if (name == rhs.name
-        && initial_state == rhs.initial_state
-        && agent_outputs == rhs.agent_outputs
-        && keepDefaultState == rhs.keepDefaultState
-        && functions.size() == rhs.functions.size()
-        && variables.size() == rhs.variables.size()
-        && states.size() == rhs.states.size()) {
-        {  // Compare functions
-                for (auto &v : functions) {
-                    auto _v = rhs.functions.find(v.first);
-                    if (_v == rhs.functions.end())
-                        return false;
-                    if (v.second->operator==(*_v->second))
-                        return false;
-                }
-        }
-        {  // Compare variables
-            for (auto &v : variables) {
-                auto _v = rhs.variables.find(v.first);
-                if (_v == rhs.variables.end())
-                    return false;
-                if (v.second.type_size != _v->second.type_size
-                    || v.second.type != _v->second.type
-                    || v.second.elements != _v->second.elements)
-                    return false;
-            }
-        }
-        {  // Compare state lists
-            if (states != rhs.states)
-                return false;
-        }
-        return true;
-    }
-    return false;
-}
-bool AgentData::operator!=(const AgentData& rhs) const {
-    return !operator==(rhs);
-}
-bool AgentFunctionData::operator==(const AgentFunctionData& rhs) const {
-    if (this == &rhs)  // They point to same object
-        return true;
-    if ((name == rhs.name)
-        && (func == rhs.func)
-        && (initial_state == rhs.initial_state)
-        && (end_state == rhs.end_state)
-        && (message_output_optional == rhs.message_output_optional)
-        && (has_agent_death == rhs.has_agent_death)) {
-        // Test weak pointers
-            {  // message_input
-                auto a = message_input.lock();
-                auto b = rhs.message_input.lock();
-                if (a && b) {
-                    if (*a != *b)
-                        return false;
-                } else if ((a && !b) || (!a && !b)) {
-                    return false;
-                }
-            }
-            {  // message_output
-                auto a = message_output.lock();
-                auto b = rhs.message_output.lock();
-                if (a && b) {
-                    if (*a != *b)
-                        return false;
-                } else if ((a && !b) || (!a && !b)) {
-                    return false;
-                }
-            }
-            {  // agent_output
-                auto a = agent_output.lock();
-                auto b = rhs.agent_output.lock();
-                if (a && b) {  // Comparing agents here is unsafe, as agent might link back to this same function, so do a reduced comparison
-                    if (a->name != b->name
-                        || a->initial_state != b->initial_state
-                        || a->agent_outputs != b->agent_outputs
-                        || a->keepDefaultState != b->keepDefaultState
-                        || a->states.size() != b->states.size()
-                        || a->functions.size() != b->functions.size()
-                        || a->variables.size() != b->variables.size()
-                        || a->states != b->states)
-                        return false;
-                    {  // Compare agent variables
-                        for (auto &v : a->variables) {
-                            auto _v = b->variables.find(v.first);
-                            if (_v == b->variables.end())
-                                return false;
-                            if (v.second.type_size != _v->second.type_size
-                                || v.second.type != _v->second.type
-                                || v.second.elements != _v->second.elements)
-                                return false;
-                        }
-                    }
-                    {  // Compare agent functions (compare name only)
-                        for (auto &v : a->functions) {
-                            auto _v = b->functions.find(v.first);
-                            if (_v == b->functions.end())
-                                return false;
-                        }
-                    }
-                } else if ((a && !b) || (!a && !b)) {
-                    return false;
-                }
-            }
-            return true;
-    }
-    return false;
-}
-bool AgentFunctionData::operator!=(const AgentFunctionData& rhs) const {
-    return !operator==(rhs);
-}
-bool LayerData::operator==(const LayerData& rhs) const {
-    if (this == &rhs)  // They point to same object
-        return true;
-    if (name == rhs.name
-        && index == rhs.index
-        && agent_functions.size() == rhs.agent_functions.size()
-        && host_functions.size() == rhs.host_functions.size()
-        && agent_functions == rhs.agent_functions
-        && host_functions == rhs.host_functions) {
-            return true;
-    }
-    return false;
-}
-bool LayerData::operator!=(const LayerData& rhs) const {
-    return !operator==(rhs);
-}
-
-bool AgentData::isOutputOnDevice() const {
-    return agent_outputs > 0;
 }
